Mmmenu#current_item now returns a hash containing curren_item's info (#more)
For example, if I built a menu with @menu.build,
I can later output current item's title in my views with
<%= @menu.current_item[:title] %>

The hash returned by the #current_item is equivalent to the hash passed by
you to the #add method while building a menu. #current_method always returns
the deepest current item, that is if you have a menu if the following structure:

    Item1:  current
      Create: /items1/new current
      Index: /items1/
      Print: /items1/print
    Item2: /items2
    Item3: /items3
    Item4: /items4

then 'Create' item will be returned.